Add a basic build workflow to the repository.
This workflow runs manually on branches and is required on
pull-requests.

Attempt building on 22.04, 24.04 for x64 and arm to demonstrate that it
isn't broken on any of those platforms.

Made the appimage build parametric so that it can be disabled during the
workflow and potentially later steps that might not want the appimage.